# documentation/swagger.yaml
openapi: 3.0.1
info:
  title: Datak
  version: 1.0.0
servers:
- url: localhost/api/v1
- url: localhost/api/v1
...
paths:
...
  /update-profile:
    post:
      security:
        - bearerAuth : []
      tags:
      - user
      summary: edit profile user
      description: .
      operationId: UpdateProfile
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: name of applicant
                image:
                  type: string
                  format: binary
                  description: image for profile
                
      responses:
        200:
          description: successful operation
          content: 
            application/json:
              schema:
                  required: 
                    - "message"
                  properties: 
                    message: 
                      type: "string"

        422:
          description: validation error
          content: 
            application/json:
              schema:
                  required: 
                    - "name"
                    - "image"
                  properties: 
                    name: 
                      type: "string" 
                    image: 
                      type: "string" 
  /event:
    get:
      security:
        - bearerAuth : []
      tags:
      - event
      summary: users event list
      description: .
      operationId: UserEvent
                
      responses:
        200:
          description: successful operation
          content: 
            application/json:
              schema:
                  required: 
                    - "message"
                    - "data"
                  properties: 
                    result: 
                      type: "string"
                    data: 
                      type: "array"
                      items: 
                        type: "object"
                        properties: 
                          id: 
                            type: "number"
                          name: 
                            type: "string"
                          onwer_id: 
                            type: "number"
                          due_date: 
                            type: "string"
                          created_at: 
                            type: "string"
                          updated_at: 
                            type: "string"
  

        401:
          description: unathorized
...
components:
...
  securitySchemes:
      bearerAuth:
        type: http
        scheme: bearer
        bearerFormat: JWT